clean up docs
