+++ opentelemetry-api/src/opentelemetry/propagators/__init__.py
@@ -12,6 +12,47 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+"""
+API for propagation of context.
+
+Example::
+
+    import flask
+    import requests
+    from opentelemetry import propagators
+
+
+    PROPAGATOR = propagators.get_global_httptextformat()
+
+
+    def get_header_from_flask_request(request, key):
+        return request.headers.get_all(key)
+
+    def set_header_into_requests_request(request: requests.Request,
+                                            key: str, value: str):
+        request.headers[key] = value
+
+    def example_route():
+        context = PROPAGATOR.extract(
+            get_header_from_flask_request,
+            flask.request
+        )
+        request_to_downstream = requests.Request(
+            "GET", "http://httpbin.org/get"
+        )
+        PROPAGATOR.inject(
+            set_header_into_requests_request,
+            request_to_downstream,
+            context=context
+        )
+        session = requests.Session()
+        session.send(request_to_downstream.prepare())
+
+
+.. _Propagation API Specification:
+    https://github.com/open-telemetry/opentelemetry-specification/blob/master/specification/api-propagators.md
+"""
+
 import typing
 
 import opentelemetry.trace as trace
@@ -28,20 +69,18 @@ def extract(
     carrier: httptextformat.HTTPTextFormatT,
     context: typing.Optional[Context] = None,
 ) -> Context:
-    """Load the parent SpanContext from values in the carrier.
-
-    Using the specified HTTPTextFormatter, the propagator will
-    extract a SpanContext from the carrier. If one is found,
-    it will be set as the parent context of the current span.
+    """ Uses the configured propagator to extract a Context from the carrier.
 
     Args:
         get_from_carrier: a function that can retrieve zero
             or more values from the carrier. In the case that
             the value does not exist, return an empty list.
         carrier: and object which contains values that are
-            used to construct a SpanContext. This object
+            used to construct a Context. This object
             must be paired with an appropriate get_from_carrier
             which understands how to extract a value from it.
+        context: an optional Context to use. Defaults to current
+            context if not set.
     """
     return get_global_httptextformat().extract(
         get_from_carrier, carrier, context
@@ -53,12 +92,7 @@ def inject(
     carrier: httptextformat.HTTPTextFormatT,
     context: typing.Optional[Context] = None,
 ) -> None:
-    """Inject values from the current context into the carrier.
-
-    inject enables the propagation of values into HTTP clients or
-    other objects which perform an HTTP request. Implementations
-    should use the set_in_carrier method to set values on the
-    carrier.
+    """ Uses the configured propagator to inject a Context into the carrier.
 
     Args:
         set_in_carrier: A setter function that can set values
@@ -66,6 +100,8 @@ def inject(
         carrier: An object that contains a representation of HTTP
             headers. Should be paired with set_in_carrier, which
             should know how to set header values on the carrier.
+        context: an optional Context to use. Defaults to current
+            context if not set.
     """
     get_global_httptextformat().inject(set_in_carrier, carrier, context)
